# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sales automation API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Business {
  id: number;
  userId: string;
  name: string;
  /** @nullable */
  serviceUrl?: string | null;
  /** @nullable */
  companyName?: string | null;
  /** @nullable */
  senderName?: string | null;
  /** @nullable */
  senderEmail?: string | null;
  /** @nullable */
  signatureHtml?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateBusinessBody {
  name: string;
  /** @nullable */
  serviceUrl?: string | null;
  /** @nullable */
  companyName?: string | null;
  /** @nullable */
  senderName?: string | null;
  /** @nullable */
  senderEmail?: string | null;
  /** @nullable */
  signatureHtml?: string | null;
}

export interface UpdateBusinessBody {
  name?: string;
  /** @nullable */
  serviceUrl?: string | null;
  /** @nullable */
  companyName?: string | null;
  /** @nullable */
  senderName?: string | null;
  /** @nullable */
  senderEmail?: string | null;
  /** @nullable */
  signatureHtml?: string | null;
}

export type LeadStatus = (typeof LeadStatus)[keyof typeof LeadStatus];

export const LeadStatus = {
  unsent: "unsent",
  sent: "sent",
  replied: "replied",
  ng: "ng",
  unsubscribed: "unsubscribed",
} as const;

export interface Lead {
  id: number;
  businessId: number;
  /** @nullable */
  companyName?: string | null;
  /** @nullable */
  websiteUrl?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  contactUrl?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  status: LeadStatus;
  /** @nullable */
  score?: number | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateLeadBodyStatus =
  (typeof CreateLeadBodyStatus)[keyof typeof CreateLeadBodyStatus];

export const CreateLeadBodyStatus = {
  unsent: "unsent",
  sent: "sent",
  replied: "replied",
  ng: "ng",
  unsubscribed: "unsubscribed",
} as const;

export interface CreateLeadBody {
  businessId: number;
  /** @nullable */
  companyName?: string | null;
  /** @nullable */
  websiteUrl?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  contactUrl?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  status?: CreateLeadBodyStatus;
  /** @nullable */
  score?: number | null;
}

export type UpdateLeadBodyStatus =
  (typeof UpdateLeadBodyStatus)[keyof typeof UpdateLeadBodyStatus];

export const UpdateLeadBodyStatus = {
  unsent: "unsent",
  sent: "sent",
  replied: "replied",
  ng: "ng",
  unsubscribed: "unsubscribed",
} as const;

export interface UpdateLeadBody {
  /** @nullable */
  companyName?: string | null;
  /** @nullable */
  websiteUrl?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  contactUrl?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  status?: UpdateLeadBodyStatus;
  /** @nullable */
  score?: number | null;
}

export interface Template {
  id: number;
  businessId: number;
  name: string;
  subjectTemplate: string;
  htmlTemplate: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateTemplateBody {
  businessId: number;
  name: string;
  subjectTemplate: string;
  htmlTemplate: string;
}

export interface UpdateTemplateBody {
  name?: string;
  subjectTemplate?: string;
  htmlTemplate?: string;
}

export interface GenerateAiEmailBody {
  leadId: number;
}

export interface GeneratedEmail {
  subject: string;
  html: string;
}

export type CampaignStatus =
  (typeof CampaignStatus)[keyof typeof CampaignStatus];

export const CampaignStatus = {
  draft: "draft",
  running: "running",
  completed: "completed",
  paused: "paused",
} as const;

export interface Campaign {
  id: number;
  businessId: number;
  templateId: number;
  name: string;
  status: CampaignStatus;
  /** @nullable */
  scheduledAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateCampaignBody {
  businessId: number;
  templateId: number;
  name: string;
  /** @nullable */
  scheduledAt?: string | null;
}

export type UpdateCampaignBodyStatus =
  (typeof UpdateCampaignBodyStatus)[keyof typeof UpdateCampaignBodyStatus];

export const UpdateCampaignBodyStatus = {
  draft: "draft",
  running: "running",
  completed: "completed",
  paused: "paused",
} as const;

export interface UpdateCampaignBody {
  name?: string;
  templateId?: number;
  status?: UpdateCampaignBodyStatus;
  /** @nullable */
  scheduledAt?: string | null;
}

export interface SendCampaignResult {
  sent: number;
  failed: number;
  skipped: number;
  errors: string[];
}

export interface SendTestEmailBody {
  toEmail: string;
  /** @nullable */
  leadId?: number | null;
}

export interface SendTestEmailResult {
  success: boolean;
  message: string;
}

export type EmailLogStatus =
  (typeof EmailLogStatus)[keyof typeof EmailLogStatus];

export const EmailLogStatus = {
  sent: "sent",
  failed: "failed",
  bounced: "bounced",
} as const;

export interface EmailLog {
  id: number;
  leadId: number;
  /** @nullable */
  campaignId?: number | null;
  subject: string;
  html: string;
  status: EmailLogStatus;
  /** @nullable */
  sentAt?: string | null;
  /** @nullable */
  error?: string | null;
  createdAt: string;
}

export interface SearchLeadsBody {
  businessId: number;
  keyword: string;
  /** @nullable */
  location?: string | null;
  maxResults?: number;
}

export interface SearchLeadsResult {
  found: number;
  saved: number;
  leads: Lead[];
}

export interface CrawlWebsiteBody {
  url: string;
  /** @nullable */
  leadId?: number | null;
}

export interface CrawlResult {
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  companyName?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  contactUrl?: string | null;
  score: number;
}

export interface DashboardStats {
  totalLeads: number;
  unsentLeads: number;
  sentLeads: number;
  repliedLeads: number;
  ngLeads: number;
  unsubscribedLeads: number;
  totalCampaigns: number;
  totalEmailsSent: number;
  replyRate: number;
}

export interface ActivityItem {
  id: number;
  type: string;
  /** @nullable */
  leadId?: number | null;
  /** @nullable */
  companyName?: string | null;
  /** @nullable */
  subject?: string | null;
  status: string;
  createdAt: string;
}

export interface UnsubscribeResult {
  success: boolean;
  message: string;
}

export type ListLeadsParams = {
  businessId?: number;
  status?: string;
};

export type ListTemplatesParams = {
  businessId?: number;
};

export type ListCampaignsParams = {
  businessId?: number;
};

export type ListEmailLogsParams = {
  businessId?: number;
  leadId?: number;
};

export type GetDashboardStatsParams = {
  businessId?: number;
};

export type GetRecentActivityParams = {
  businessId?: number;
  limit?: number;
};