# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sales automation API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /businesses:
    get:
      operationId: listBusinesses
      tags: [businesses]
      summary: List all businesses for current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Business"
    post:
      operationId: createBusiness
      tags: [businesses]
      summary: Create a new business
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBusinessBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
  /businesses/{id}:
    get:
      operationId: getBusiness
      tags: [businesses]
      summary: Get a business by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
        "404":
          description: Not found
    patch:
      operationId: updateBusiness
      tags: [businesses]
      summary: Update a business
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBusinessBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
    delete:
      operationId: deleteBusiness
      tags: [businesses]
      summary: Delete a business
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Leads ────────────────────────────────────────────────────────────────────
...
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List campaigns for a business
      parameters:
        - in: query
          name: businessId
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get a campaign
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update a campaign
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCampaignBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /campaigns/{id}/send:
    post:
      operationId: sendCampaign
      tags: [campaigns]
      summary: Send emails for this campaign to unsent leads
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Send result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendCampaignResult"
  /campaigns/{id}/send-test:
    post:
      operationId: sendTestEmail
      tags: [campaigns]
      summary: Send a test email to a specific address
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendTestEmailBody"
      responses:
        "200":
          description: Test sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendTestEmailResult"

  # ── Email Logs ───────────────────────────────────────────────────────────────
...
  /unsubscribe/{token}:
    get:
      operationId: unsubscribeLead
      tags: [email-logs]
      summary: Process unsubscribe via token link
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Unsubscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsubscribeResult"
components:
  schemas:
...
    Business:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        serviceUrl:
          type: ["string", "null"]
        companyName:
          type: ["string", "null"]
        senderName:
          type: ["string", "null"]
        senderEmail:
          type: ["string", "null"]
        signatureHtml:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - name
        - createdAt
        - updatedAt
    CreateBusinessBody:
      type: object
      properties:
        name:
          type: string
        serviceUrl:
          type: ["string", "null"]
        companyName:
          type: ["string", "null"]
        senderName:
          type: ["string", "null"]
        senderEmail:
          type: ["string", "null"]
        signatureHtml:
          type: ["string", "null"]
      required:
        - name
    UpdateBusinessBody:
      type: object
      properties:
        name:
          type: string
        serviceUrl:
          type: ["string", "null"]
        companyName:
          type: ["string", "null"]
        senderName:
          type: ["string", "null"]
        senderEmail:
          type: ["string", "null"]
        signatureHtml:
          type: ["string", "null"]

    # ── Lead ────────────────────────────────────────────────────────────────────
...
    Campaign:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        templateId:
          type: integer
        name:
          type: string
        status:
          type: string
          enum: [draft, running, completed, paused]
        scheduledAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - businessId
        - templateId
        - name
        - status
        - createdAt
        - updatedAt
    CreateCampaignBody:
      type: object
      properties:
        businessId:
          type: integer
        templateId:
          type: integer
        name:
          type: string
        scheduledAt:
          type: ["string", "null"]
      required:
        - businessId
        - templateId
        - name
    UpdateCampaignBody:
      type: object
      properties:
        name:
          type: string
        templateId:
          type: integer
        status:
          type: string
          enum: [draft, running, completed, paused]
        scheduledAt:
          type: ["string", "null"]
    SendCampaignResult:
      type: object
      properties:
        sent:
          type: integer
        failed:
          type: integer
        skipped:
          type: integer
        errors:
          type: array
          items:
            type: string
      required:
        - sent
        - failed
        - skipped
        - errors
    SendTestEmailBody:
      type: object
      properties:
        toEmail:
          type: string
        leadId:
          type: ["integer", "null"]
      required:
        - toEmail
    SendTestEmailResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    # ── Email Log ────────────────────────────────────────────────────────────────
...
    UnsubscribeResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message